Acoustic signals

+++ arduino/libraries/TimeCube/src/TimeCube.h
@@ -11,7 +11,7 @@ class TimeCube {
 private:
   static const unsigned long BLINK_INTERVAL_ms = 200;
   static const unsigned long BLINK_PITCH = NOTE_A3;
-  static const unsigned long BLINK_PIN = 8;
+  static const int BLINK_PIN = 8;
   static const unsigned long BLINKING_DURATION_ms = 1000 * 60;
 
   unsigned long _now;
